ServerPython/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
cffi==2.0.0
cryptography==45.0.7
Django==5.2.3
grpcio==1.75.1
grpcio-tools==1.75.1
gunicorn==25.3.0
kafka-python==2.2.15
packaging==26.2
protobuf==6.32.1
psycopg==3.2.10
pycparser==2.23
PyMySQL==1.1.1
python-dotenv
pytz==2025.2
schedule==1.2.2
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions
mysql-connector-python
confluent-kafka
mysql-connector-python==9.1.0


